Add files via upload
On the first try I did, I used a NN and a Linear Regression to predict the pressure. As a first observation, the LM predicts much better the output (MAE  = 3.911 for lm vs. MAE =  11.20 for NN). 
Also, I used two ways to create the train/test split: create random splits using train_test_split and assign manually the first 70% of the data to train and the rest to test ( the reason I used the last one too is that we track breaths in different stages ).